+++ lib/recurly/account.rb
@@ -18,10 +18,10 @@ module Recurly
     end
 
     # initialize associations
-    def initialize(attributes = {})
+    def initialize(attributes = {}, persisted = false)
       attributes = attributes.with_indifferent_access
       attributes[:billing_info] ||= {}
-      super(attributes)
+      super
     end
 
     attr_accessor :account_code_was